+++ changelog/RELEASE-NOTES.md
@@ -1,1 +1,28 @@
+## 5 New Themes
+
+From the Quintessential Quintuplets series:
+
+- Nakano Nino (Dark)
+- Nakano Miku (Dark)
+
+From the Lord El-Melloi II Case Files series:
+
+- Gray (Dark)
+
+From the Daily Life with a Monster girl series:
+
+- Miia (Dark)
+
+Addition to Miss Kobayashi's Dragon Maid:
+
+- Tohru (Light)
+
+## Other stuff
+
+- Added a secondary sticker for [Hatsune Miku](https://github.com/doki-theme/doki-master-theme/issues/62)!
+- Grouped all the Type-Moon products under `Type Moon` (eg: Fate)
 - Fixed poor offline user experience. No longer propagating exceptions when offline.
+
+
+  ![v15 Girls](https://doki.assets.unthrottled.io/misc/v15_girls_smol.png)
+